requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.5.0
blinker==1.4
certifi==2024.2.2
click==8.1.7
clickhouse-connect==0.7.8
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
orjson==3.10.3
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==5.4.1
rich==13.7.1
SecretStorage==3.3.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
systemd-python==234
typer==0.12.3
typing_extensions==4.11.0
ubuntu-advantage-tools==8001
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.21.0
websockets==12.0
wsproto==1.0.0
zipp==1.0.0
zstandard==0.22.0
